refactor: move tab bar from manage page to app header
Extract tab bar UI and logic into a new HeaderTabs component and render
it in the app header's middle section, aligned to the sidebar's right
edge via var(--sidebar-width). This keeps tabs visible at the top of the
window and simplifies the manage page to only render content.
